Fix image mode compatibility issues in WSI processing
This commit fixes the "image has wrong mode" error that occurred during artifact detection by ensuring all PIL Images are in compatible modes before resize/blend operations.

Changes:
- wsi_maps.py: Convert OpenSlide thumbnail and heatmap to RGB before resize/blend
- wsi_tis_detect.py: Convert tissue mask to uint8 and OpenSlide thumbnail to RGB
- main.py: Convert loaded tissue mask to grayscale (L) mode before resize
- .gitignore: Add __pycache__, compiled Python files, and models directories

The root cause was OpenSlide returning RGBA images and tissue masks being saved as int8, which are incompatible with LANCZOS resampling and cv2.addWeighted operations.

+++ 01_WSI_inference_OPENSLIDE_QC/wsi_tis_detect.py
@@ -82,7 +82,7 @@ for slide_name in slide_names:
         mpp = round(float(slide.properties["openslide.mpp-x"]), 4)
         reduction_factor = MPP_MODEL_TD / mpp
 
-        image_or = slide.get_thumbnail((w_l0 // reduction_factor, h_l0 // reduction_factor))
+        image_or = slide.get_thumbnail((w_l0 // reduction_factor, h_l0 // reduction_factor)).convert('RGB')
         image_or.save(tis_det_dir_thumb + slide_name + ".jpg", quality = 80)
 
         '''
@@ -151,7 +151,7 @@ for slide_name in slide_names:
                 end_image = np.concatenate((end_image, temp_image), axis=0)
                 end_image_class_map = np.concatenate((end_image_class_map, temp_image_class_map), axis=0)
 
-        Image.fromarray(end_image).save(os.path.join(tis_det_dir_mask, slide_name + '_MASK.png'))
+        Image.fromarray(end_image.astype(np.uint8)).save(os.path.join(tis_det_dir_mask, slide_name + '_MASK.png'))
         Image.fromarray(end_image_class_map).save(os.path.join(tis_det_dir_mask_col, slide_name + '_MASK_COL.png'))
         overlay = cv2.addWeighted(np.array(image), OVER_IMAGE, end_image_class_map, OVER_MASK, 0)
         overlay = Image.fromarray(overlay)
